requirements.txt
python-binance
ta
pandas
requests